atualização compactador
Juntei o compactador e o descompactador em um me
