+++ src/Main.java
@@ -2,10 +2,12 @@
         public static void main(String[] args) {
 
             try{
-                Compactar compac = new Compactar();
-                Descompactar desc = new Descompactar();
-                compac.compactar("C:\\Users\\adria\\Documents\\GitHub\\estruturaDeDadosII-final\\src\\WhatsApp Image 2022-03-11 at 13.09.31.jpeg");
-                desc.extracao("C:\\Users\\adria\\Documents\\GitHub\\estruturaDeDadosII-final\\src\\WhatsApp Image 2022-03-11 at 13.09.31.jpeg.compress");
+                Compactador c = new Compactador();
+                Compactador.Compressor comp = c.new Compressor(); 
+                Compactador.Descompressor descomp = c.new Descompressor();
+                
+                comp.compress("C:\\Users\\adria\\Documents\\GitHub\\estruturaDeDadosII-final\\src\\WhatsApp Image 2022-03-11 at 13.09.31.jpeg");
+                descomp.unzip("C:\\Users\\adria\\Documents\\GitHub\\estruturaDeDadosII-final\\src\\WhatsApp Image 2022-03-11 at 13.09.31.jpeg.compress");
             }
             catch (Exception e) {
                  throw new RuntimeException(e);
